added logic to ensure backup removal only runs for local storage option

+++ nsbackup.sh
@@ -269,8 +269,11 @@ $logmsg "Info: Backup complete"
 
 done
 
-echo -e "\e[92mInfo: Removing backups older than $keepdays days...\e[39m"
-$logmsg "Info: Removing older backups..."
-find ${backup_path} -type f -mtime +${days} -exec rm -f {} \;
-echo -e "\e[92mInfo: Removal complete\e[39m\n"
-$logmsg "Info: Removal complete"
+if [ "$storage" == "local" ]
+then
+  echo -e "\e[92mInfo: Removing backups older than $keepdays days...\e[39m"
+  $logmsg "Info: Removing older backups..."
+  find ${backup_path} -type f -mtime +${days} -exec rm -f {} \;
+  echo -e "\e[92mInfo: Removal complete\e[39m\n"
+  $logmsg "Info: Removal complete"
+fi
